skin_classifier/requirements.txt
torch
torchvision
streamlit
numpy
opencv-python
Pillow
matplotlib
